Move Svg handler to images folder.
Add an arrow svg image for ObjectInspector
