+++ packages/devtools-reps/src/shared/images/Svg.js
@@ -6,6 +6,7 @@ const React = require("react");
 const InlineSVG = require("svg-inline-react");
 
 const svg = {
+  "arrow": require("./arrow.svg"),
   "open-inspector": require("./open-inspector.svg"),
 };
